Introduce a governance model file
