+++ src/canisters/governance/model.ts
@@ -1,0 +1,8 @@
+
+type NeuronId = bigint;
+
+export type KnownNeuron = {
+  id: NeuronId,
+  name: string,
+  description: string | undefined
+};
